squash! Set InfoRequest's law_used_readable_data in a method not a constant
Move InfoRequest's LAW_USED_READABLE_DATA to LawUsed.readable_data

The default locale translations are mistakenly always returned instead of
the translations for the requested locale if the translations are assigned
to a constant. Moving the Hash to an included module solves the problem
without adding a random method definition to the top of the InfoRequest
model file.